backend/requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
pandas
tabulate
pytest
httpx
python-dotenv
